Sarcasm Detection using RNN
The model developed in the project classifies news that whether it is sarcastic or not. Dataset consists of labeled news headlines collected from two news websites, The Onion and Huffpost published on Kaggle website. After preprocessing steps involved handling missing values, stop words removal, tokenizing, stemming and lemmatization, LSTM is used to develop a Recurrent Neural Network which is then trained on the preprocessed data for the purpose.